eda/reddit-sentimental-analysis/requirements.txt
datetime
pandas
praw
pyyaml